Misc cleanup in he_core
- Fix natstepfilter entries
- Fix tuple bugs on MSVC
- Fix some casts in win32 thread library
- Sort and cleanup some includes
- Update a few out-of-date tests

+++ plugins/core/src/win32/thread.win32.cpp
@@ -124,7 +124,7 @@ namespace he
 
     Result TlsValue::Create(Pfn_TlsDestructor destroy) noexcept
     {
-        m_id = ::FlsAlloc(destroy);
+        m_id = static_cast<uintptr_t>(::FlsAlloc(destroy));
         return m_id == InvalidId ? Result::FromLastError() : Result::Success;
     }
 
@@ -132,21 +132,21 @@ namespace he
     {
         if (m_id != InvalidId)
         {
-            ::FlsFree(m_id);
+            ::FlsFree(static_cast<DWORD>(m_id));
             m_id = InvalidId;
         }
     }
 
     void* TlsValue::Get() const noexcept
     {
-        return m_id == InvalidId ? nullptr : ::FlsGetValue(m_id);
+        return m_id == InvalidId ? nullptr : ::FlsGetValue(static_cast<DWORD>(m_id));
     }
 
     void TlsValue::Set(void* value) noexcept
     {
         if (m_id != InvalidId)
         {
-            ::FlsSetValue(m_id, value);
+            ::FlsSetValue(static_cast<DWORD>(m_id), value);
         }
     }
 }
